Add files via upload
Images were loading and crashing. What I did was remove mouseover effects and enabled the "click to view" images to be only the back view of garments. This way we progressed our momentum of launching the site and can revisit the goal view of mouseover effects and "click to view" to be a zoomed or larger in the image once a "lazy load" effect is applied. Funny enough, the shoes page has the most images of all the pages yet loads effortlessly and functions properly with mouseover effects and such.